Brind tslr and add eligibility checking in line
Updates tslr eligibility checking to match additional payments naming
and calling conventions
