+++ app/models/journeys/teacher_student_loan_reimbursement/eligibility_checker.rb
@@ -1,23 +1,6 @@
 module Journeys
   module TeacherStudentLoanReimbursement
-    class EligibilityChecker
-      attr_reader :journey_session
-
-      delegate_missing_to :checker
-
-      def initialize(journey_session:)
-        @journey_session = journey_session
-      end
-
-      # FIXME this could be shared with the additional payments eligibility
-      # checker, just make a journey eligibility checker
-      private
-
-      def checker
-        @checker ||= Policies::StudentLoans::EligibilityChecker.new(
-          journey_session.answers
-        )
-      end
+    class EligibilityChecker < Journeys::EligibilityChecker
     end
   end
 end
